Versión final 3.0

+++ src/logica/borrado.py
@@ -6,6 +6,9 @@ from src.modelo.declarative_base import Session, engine, Base
 if __name__ == '__main__':
     session = Session()
     cancion2 = session.query(Cancion).get(2)
-    session.delete(cancion2)
-    session.commit()
+    if cancion2 != None:
+        session.delete(cancion2)
+        session.commit()
+    else:
+        print("El registro no existe")
     session.close()
